feat: add functionality to insert data

+++ src/products/entities/product.entity.ts
@@ -8,7 +8,7 @@ export class Product {
   @Column('text', { unique: true })
   title: string;
 
-  @Column('numeric', { default: 0 })
+  @Column('float', { default: 0 })
   price: number;
 
   @Column({ type: 'text', nullable: true })
